[Silver II] Title: 이항 계수 2, Time: 4 ms, Memory: 5936 KB -BaekjoonHub

+++ 백준/Silver/11051. 이항 계수 2/이항 계수 2.cc
@@ -5,7 +5,7 @@ using namespace std;
 typedef long long ll;
 
 const int MOD = 10007;
-ll dp[1001][1001];
+int dp[1001][1001];
 
 int main() {
     int N, K;
